refactor(ai-writeback): split service into constants, types, templates (part-3)
